Changed all rem units to global sizing variables. Renamed css file to dashboardHeader. Fixed click funcionality using new background banner wrapper.
