+++ frontend/compositions/dashboardHeader/index.tsx
@@ -4,8 +4,7 @@ import LogoMobile from './assets/LogoMobile.svg'
 import NPDCLogo from './assets/NPDCLogo.svg'
 import HamburgerMenu from './assets/Vector.svg'
 import DesktopNav from './desktopNav'
-import styles from './header.module.css'
-
+import styles from './dashboardHeader.module.css'
 
 export default function DashboardHeader() {
   const {
@@ -27,24 +26,24 @@ export default function DashboardHeader() {
   return (
     <header className={wrapper}>
       <div className={backgroundBanner}>
-      <div className={leftHeader}>
-        <div className={logoContainer}>
-          {logo}
+        <div className={leftHeader}>
+          <div className={logoContainer}>
+            {logo}
+          </div>
+          <div className={titleContainer}>
+            <h2>{title}</h2>
+            <p>The national index of police incidents</p>
+          </div>
         </div>
-        <div className={titleContainer}>
-          <h2>{title}</h2>
-          <p>The national index of police incidents</p>
-        </div>
-      </div>
 
-      <div className={rightHeader}>
-        <nav className={nav}>
-          {navBar}
-        </nav>
-        <button className={button} type="button">
-          <p>DONATE</p>
-        </button>
-      </div>
+        <div className={rightHeader}>
+          <nav className={nav}>
+            {navBar}
+          </nav>
+          <button className={button} type="button">
+            <p>DONATE</p>
+          </button>
+        </div>
       </div>
     </header>
   )
